+++ CHANGELOG.md
@@ -11,6 +11,11 @@
 * Added `SearchFlip::Connection#get_cluster_settings` and
   `#update_cluster_settings`
 
+## v3.5.0
+
+* Add `SearchFlip::Criteria#http_timeout` to allow specifying timeouts on
+  a query level
+
 ## v3.4.0
 
 * Expose `Http#timeout` via `SearchFlip::HTTPClient`